refactor mcp rule warning diagnostics
